crimson/os/seastore/rbm: move RBMDevice related functions to rbm_device.cc
